requirements.txt
python-dotenv
requests
pandas
mysqlclient